src/requirements.txt
numpy==1.14.0
scipy==0.19.0
certifi==2020.6.20
dill==0.2.7.1
Keras-Preprocessing==1.1.0
llvmlite==0.33.0
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mlxtend==0.17.3
mock==4.0.2
numba==0.41.0
pandas==1.1.3
Pillow==5.0.0
python-dateutil==2.8.1
python-utils==2.4.0
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
simpy==4.0.1
six==1.15.0
sympy==1.6.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
toposort==1.5
Werkzeug==1.0.1
xxhash==1.0.1
